cross_domain_evaluation/implementation/gutget/requirements.txt
acme==1.12.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certbot==1.12.0
certbot-nginx==1.10.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
ConfigArgParse==1.2.3
configobj==5.0.6
croniter==0.3.34
cryptography==3.3.2
dbus-python==1.2.16
distro==1.9.0
exceptiongroup==1.2.2
flask==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.18.1
httpx==0.28.1
idna==2.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
josepy==1.2.0
MarkupSafe==3.0.2
msgpack==1.0.0
natsort==7.1.0
openai==1.63.2
packaging==24.2
parsedatetime==2.6
psutil==5.8.0
pycryptodomex==3.9.7
pydantic==2.10.6
pydantic-core==2.27.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-dateutil==2.8.1
python-debian==0.1.39
python-debianbts==3.1.0
python-dotenv==1.0.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
requests-toolbelt==0.9.1
salt==3002.6
six==1.16.0
sniffio==1.3.1
urllib3==1.26.5
werkzeug==3.1.3
zipp==3.21.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0